requirements.txt
pandas==2.2.3
numpy==2.0.2
matplotlib==3.9.4
seaborn==0.13.2
python-telegram-bot==22
yfinance==0.2.55
python-dateutil==2.9.0